move admin stuff to own component
